Introduce closed workload model support, close #1647
Motivation:

Gatling currently only support open workload model.
It would be interesting to support closed workload model too as it
sometimes matches best actual system’s behavior.

Modifications:

* Introduce a new DSL for closed workload model. `inject` can now be
passed either open steps or closed ones (but not a mix)
* Disable total users for closed model as there’s no way to know
beforehand
* Improve documentation

Result:

Gatling now supports closed workload model

+++ gatling-metrics/src/main/scala/io/gatling/metrics/types/UserBreakdownBuffer.scala
@@ -19,12 +19,12 @@ package io.gatling.metrics.types
 import io.gatling.core.stats.message.{ End, Start }
 import io.gatling.core.stats.writer.UserMessage
 
-private[metrics] class UserBreakdownBuffer(val totalUserEstimate: Int) {
+private[metrics] class UserBreakdownBuffer(val totalUserEstimate: Long) {
 
-  private var previousActive = 0
-  private var previousEnd = 0
-  private var thisStart = 0
-  private var thisEnd = 0
+  private var previousActive = 0L
+  private var previousEnd = 0L
+  private var thisStart = 0L
+  private var thisEnd = 0L
 
   private var start = 0
   private var end = 0
@@ -52,4 +52,4 @@ private[metrics] class UserBreakdownBuffer(val totalUserEstimate: Int) {
   }
 }
 
-private[metrics] case class UserBreakdown(active: Int, waiting: Int, done: Int)
+private[metrics] case class UserBreakdown(active: Long, waiting: Long, done: Long)
